ml/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Faker==37.0.0
fastapi==0.115.11
fonttools==4.56.0
google-auth==2.38.0
google-genai==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
mlxtend==0.23.4
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
python-dotenv==1.0.1